Separate Funding by Region
