+++ MapSci.html
@@ -1,0 +1,59 @@
+<!DOCTYPE html>
+<html lang="en">
+
+    <head>
+        <meta charset="utf-8">
+        <title>Research Funding in the USA</title>
+        <script src="http://d3js.org/d3.v3.min.js" charset="utf-8"></script>
+
+        <style>
+            body {
+                background-color: white;
+            }
+        </style>
+
+    </head>
+
+    <body>
+
+        <script type="text/javascript" src="Map-Sci.d3.js">
+        </script>
+
+    </body>
+
+</html>
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
